Draft: Support for Lenovo ThinkSystem SR650 and SR670.
REVIEWER NOTE: Everything has moved, so it is probably easier to review
this by checking out the branch, rather than looking at the diff.

Now that we have multiple vendors I moved some things around.

Test it:
 - `cd example`
 - `cargo run -- -H 10.153.145.92 -U Egxadmin -P '<removed>' -V Lenovo
   -c get_power_state` (ask me for pw)

The public methods in src/lib.rs are now behind a `Redfish` trait with eight
methods - two diagnostic ones and six that Carbide needs. This trait is
the only part of the package that Carbide will need to know about.

You get an implementation like this:
```
    let redfish = libredfish::new(
	Vendor::Lenovo,
	NetworkConfig{endpoint: "10.153.145.92", Default::default()..},
    )?;
```
See `example/src/main.rs` for usage examples.

The redfish data model is in model/ with the OEM specific parts in
model/oem/. The model fixtures moved from tests/ to `model/testdata`.

Implementations of the `Redfish` trait rely on a `RedfishStandard` object for
the parts that actually follow the standard and are therefore common
between providers (not a whole lot unfortunately). The custom parts are
in a file named after the provider, e.g. `src/dell.rs`.

The `RedfishStandard` object has a `Network` object which wraps and
hides `reqwest`.

There are a bunch of things in here that Carbide doesn't use, but might
later (e.g. thermal).