feat: Allow creating a standard client with custom headers
create_client_with_custom_headers enables carbide to talk to all BMC's
through a proxy, using custom headers to tell which BMC we actually want
to talk to, but we're unable to do so for certain code paths which
create a "standard" client (ie. with no HTTP calls). This fills in that
gap.