feat: add functionality to enable/disable IPMI over LAN for a server.
feat: add functionality to enable/disable IPMI over LAN for a server. Also let clients retrieve the status of this property

We need this functionality because IPMI over LAN is disabled for a server after a factory reset. Forge will have to proactively enable this property  in the state machine.