fix: modify how we get/set KCS enabled on Lenovos
```
fix: modify how we set whether KCS is enabled on a Lenovo to accomodate the value being either a boolean or a String in the bios attributes. we have observed that this value is a boolean on 650v2/670v2s but is a string (enabled/disabled) on 655v3/675v3s
```

```
fix: query the manager structure to find whether KCS is enabled/disabled on Lenovos. The manager structure has custom deserialization to handle the KCS enabled value being both a boolen or a string value (Enabled/Disabled)
```