feat: do not issue force restart against SR 675 V3 OVX servers
feat: we have observed that issuing a force restart against SR 675 V3 OVX servers can get them into a hung state where the power is on, but the machine hangs at UEFI. Lenovo has provided GracefulRestart as a work-around for this. So, we are modifying the lenovo implementation for power control. We will not issue a force restart against SR 675 V3 OVX servers.