fix: issue work-around provided by Lenovo instead of issuing a ForceRestart
Issue work-around provided by Lenovo instead of issuing a ForceRestart to the SR 675s in Azure Phase 2B. We tried issuing a GracefulRestart, but that is not a reliable mechanism of rebooting the server: the servers will still get stuck in a state where we have to AC power cycle them.

This is a specific workaround for the SKU we are receiving for GPU nodes in Azure Phase 2B. Lenovo will be providing us with a newer firmware version that will fix the ForceRestart issue on their end in 4 weeks. Adding a sleep for 10 seconds is not ideal in the libredfish package, but it is significantly less than the default timeout of 60 seconds.