fix: Return HTTPStatusCode if we can't parse error json
Currently if the redfish server returns a 403 Forbidden with XML in the
response (as happens in DGX H100 BMC's), we try to see if the response
hints that a password needs to be changed, to return a custom error in
that case. But if the JSON doesn't parse (because it's really XML in
this case), we return a JsonDeserializeError, which isn't right... if
the response was a failed HTTP code, the caller needs to know that.

This was ultimately causing forge to reset BMC's every time we got a
forbidden response, when normally getting an HTTPStatusCode error would
prevent us from doing that.

Fix is to ignore any parsing errors we get when trying to guess if a
forbidden response is due to needing a password change, and just
propagate it as a forbidden response.