feat: add BIOS factory reset functionality + fix BMC factory reset for HPEs
JIRAs: [FORGE-5231](https://jirasw.nvidia.com/browse/FORGE-5231) | [FORGE-5964](https://jirasw.nvidia.com/browse/FORGE-5964)

This MR uses the following path(s) to reset the BIOS to factory defaults. The systems require a reboot for the new settings to apply.

* Dell, HPE, SuperMicro
  * `POST redfish/v1/Systems/{}/Bios/Actions/Bios.ResetBios`
* Lenovo
  * `POST redfish/v1/Systems/{}/Bios/Actions/Bios.ResetBios -d '{"ResetType": "Reset"}'`
* Vikings
  * `POST UpdateService/Actions/Oem/NvidiaUpdateService.ClearNVRAM`
* DPUs
  * `PATCH redfish/v1/Systems/{}/Bios/Settings -d '{"Attributes":{"ResetEfiVars": true}}'`

It also adds the proper path to factory reset the iLO for HPE machines as the standard URL threw errors.

This MR is part of a larger effort to incorporate resetting the BIOS/BMC to factory defaults when force-deleting a machine.

* Corresponding carbide MR: https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/3952