fix: query nic mode from the BIOS attributes on BF3s only if the BF3 is actually in NIC mode. Otherwise, use the OEM specific extension.
The OEM specific extension is the most reliable way to determine a BF3's NIC Mode. However, we know that the OEM extension will time out if a BF3 is in NIC mode and has a BMC FW of BF-24.04-5. Work around this by checking if a DPU at BF-24.04-5 is in NIC mode. Only query the OEM specific extension if its not in NIC mode to verify that it is in DPU mode.