fix: use if-match header when setting password on Vikings via the standard redfish client
After resetting a Vikings' BMC to factory-defaults and attempting to reingest it, the following error was encountered:

```
level=ERROR span_id=0xb1250e0eb8fe45fa msg="Failed to create Redfish client while setting BMC password for vendor AMI
 (initialize = true, bmc_ip = 10.217.155.75:443): RedfishError(HTTPErrorCode { url: 
\"https://10.217.155.75:443/redfish/v1/Systems/\", status_code: 401, response_body: \"{\\\"error\\\":
{\\\"@Message.ExtendedInfo\\\":[{\\\"@odata.type\\\":\\\"#Message.v1_0_8.Message\\\",\\\"Message\\\":\\\"While 
attempting to establish a connection to /redfish/v1/Systems/, the service was denied access.\\\",\\\"MessageArgs\\\":
[\\\"/redfish/v1/Systems/\\\"],\\\"MessageId\\\":\\\"Security.1.0.AccessDenied\\\",\\\"Resolution\\\":\\\"Attempt to 
ensure that the URI is correct and that the service has the appropriate credentials.\\\",\\\"Severity\\\":\\\"Critical\\\"}],\\\"code\\\":\\\"Security.1.0.AccessDenied\\\",\\\"message\\\":\\\
"While attempting to establish a connection to /redfish/v1/Systems/, the service was denied access.\\\"}}\" })" 
location="api/src/site_explorer/redfish.rs:137"
```

This meant that setting `initialize = true` when attempting to create a Viking-specific redfish client in the site explorer code path for changing the BMC password failed. This was not seen before, and could possibly be a side effect of upgrading the Viking BIOS to 1.6.7.

This MR moves `patch_with_if_match()` to network.rs so it can be used by any vendor and thereby the standard redfish implementation. This was tested on the stuck machine above and the new standard redfish call to change the password worked successfully.

Corresponding carbide MR: https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/4137