fix: wait for the lifecycle controller to become ready before we interact with the BOSS
When a tenant releases an instance that has a BOSS controller in the server, the state machine needs to decommission the boss controller and creating a volume on top of it as part of HostCleanup. We have sometimes seen errors when initiating bios config jobs to interact with the BOSS controller. Dell asked us to query the lifecycle controller status prior to creating the bios config jobs.

 https://github.com/dell/idrac-Redfish-Scripting/issues/323